- brats 2019 (seg)

+++ datasets.py
@@ -304,7 +304,7 @@ class GenericSegmentationCroppingDataset(Dataset):
         inputs = np.expand_dims(case[slice_i], 0)
 
         if self.labels is not None:
-            labels_patch = get_image(self.labels[case_idx])
+            labels_patch = get_image(self.labels[case_idx]).astype(long)
             labels = np.expand_dims(labels_patch, 0)
 
             return inputs, labels
